src/Pipeline1/requirements.txt
python-dotenv
pyaudio
transformers
datasets
torch
soundfile
datasets